refactor(github): isolate GitHub domain into dedicated layer
Extract all GitHub API concerns into scripts/github/ following a
ports-and-adapters approach: shared HTTP client in githubApiClient.ts,
GraphQL queries in queries.ts, and renamed fetchers (fetchUserStats,
fetchUserLanguages) fixing the fetchUserData export name mismatch.
Also fix Object.values() misuse on arrays and add request timeout.
